src/main/java/com/rest/springbootemployee/EmployeeService.java
package com.rest.springbootemployee;

import org.springframework.stereotype.Service;

import java.util.List;
@Service
public class EmployeeService {

    private  EmployeeRepository employeeRepository;

    public  EmployeeService(EmployeeRepository employeeRepository)
    {
        this.employeeRepository =employeeRepository;
    }


    public List<Employee> findAll(){
        return  employeeRepository.findAll();
    }

    public Employee update(Integer id, Employee employee) {
        Employee existingEmployee = employeeRepository.findById(id);
        if (employee.getAge() != null) {
            existingEmployee.setAge(employee.getAge());
        }
        if (employee.getSalary() != null) {
            existingEmployee.setSalary(employee.getSalary());
        }
        return existingEmployee;
    }
}
